src/app/pipes/product-filter.pipe.ts
import { Pipe, PipeTransform } from '@angular/core';
import { Product } from '../models/product.model';

@Pipe({
  name: 'productFilter'
})
export class ProductFilterPipe implements PipeTransform {
  transform(products: Product[], searchText: string): Product[] {
    if (!products || !searchText) {
      return products;
    }

    searchText = searchText.toLowerCase();

    return products.filter(product => {
      return (
        product.name.toLowerCase().includes(searchText) ||
        product.description.toLowerCase().includes(searchText)
      );
    });
  }
}

